faster rcnn shell config

+++ mmdet/datasets/xml_style.py
@@ -45,10 +45,10 @@ class XMLDataset(CustomDataset):
             difficult = int(obj.find('difficult').text)
             bnd_box = obj.find('bndbox')
             bbox = [
-                int(bnd_box.find('xmin').text),
-                int(bnd_box.find('ymin').text),
-                int(bnd_box.find('xmax').text),
-                int(bnd_box.find('ymax').text)
+                float(bnd_box.find('xmin').text),
+                float(bnd_box.find('ymin').text),
+                float(bnd_box.find('xmax').text),
+                float(bnd_box.find('ymax').text)
             ]
             if difficult:
                 bboxes_ignore.append(bbox)
